MAINT: stats: rename _bootstrap to _resampling
