+++ scipy/stats/__init__.py
@@ -476,8 +476,8 @@ from . import qmc
 from ._multivariate import *
 from . import contingency
 from .contingency import chi2_contingency
-from ._bootstrap import (bootstrap, BootstrapDegenerateDistributionWarning,
-                         monte_carlo_test)
+from ._resampling import (bootstrap, BootstrapDegenerateDistributionWarning,
+                          monte_carlo_test)
 from ._entropy import *
 from ._hypotests import *
 from ._rvs_sampling import rvs_ratio_uniforms, NumericalInverseHermite  # noqa
