+++ scipy/stats/_resampling.py
