+++ scipy/stats/meson.build
@@ -186,7 +186,6 @@ py3.install_sources([
     '_axis_nan_policy.py',
     '_binned_statistic.py',
     '_binomtest.py',
-    '_bootstrap.py',
     '_common.py',
     '_constants.py',
     '_continuous_distns.py',
@@ -208,6 +207,7 @@ py3.install_sources([
     '_page_trend_test.py',
     '_qmc.py',
     '_relative_risk.py',
+    '_resampling.py',
     '_result_classes.py',
     '_rvs_sampling.py',
     '_stats_mstats_common.py',
