+++ scipy/stats/tests/meson.build
@@ -4,7 +4,6 @@ py3.install_sources([
     'studentized_range_mpmath_ref.py',
     'test_axis_nan_policy.py',
     'test_binned_statistic.py',
-    'test_bootstrap.py',
     'test_contingency.py',
     'test_continuous_basic.py',
     'test_crosstab.py',
@@ -22,6 +21,7 @@ py3.install_sources([
     'test_qmc.py',
     'test_rank.py',
     'test_relative_risk.py',
+    'test_resampling.py',
     'test_sampling.py',
     'test_stats.py',
     'test_tukeylambda_stats.py',
